switch output dtypes

+++ transform/reta/src/transform_reta.py
@@ -40,7 +40,7 @@ DTYPES = {
     "month": str,
     "card": str,
     "category": str,
-    "value": float,
+    "value": int,
 }
